backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==13.0.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# ML/NLP - Use latest compatible versions
torch>=2.2.0
transformers>=4.37.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0

# NLP Specific
vaderSentiment==3.3.2
langdetect==1.0.9

# ML Models
xgboost>=2.0.0

# Data Processing
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate Limiting
slowapi==0.1.9

# Async
aiohttp>=3.9.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Monitoring
structlog==24.1.0

